Move login fragment to fragment package
